LayoutTests/media/track/media-video-audio-track-expected.txt


1) Disable and enable an audio track in a run loop will enable the audio track.
2) Deselect and select a video track in a run loop will select the video track.

RUN(mediaElement.src = findMediaFile("video", "../content/test"))

EVENT(canplaythrough)
EXPECTED (video.audioTracks.length == '1') OK
EXPECTED (video.audioTracks[0].enabled == 'true') OK
EXPECTED (video.videoTracks.length == '1') OK
EXPECTED (video.videoTracks[0].selected == 'true') OK
RUN(video.audioTracks[0].enabled = false)
RUN(video.audioTracks[0].enabled = true)
RUN(video.videoTracks[0].selected = false)
RUN(video.videoTracks[0].selected = true)
END OF TEST
